requirements.txt
ipywidgets
ipykernel
python-dotenv
openai
fastapi
uvicorn
